# docs/swagger.yaml
definitions:
  http.Car:
    properties:
      hsn:
        type: string
      manufacturer:
        type: string
      name:
        type: string
      tsn:
        type: string
    type: object
  models.Car:
    properties:
      hsn:
        example: "0001"
        type: string
      manufacturer_name:
        example: ADLERWERKE
        type: string
      name:
        example: M 100
        type: string
      tsn:
        example: "096"
        type: string
    type: object
info:
  contact: {}
paths:
  /api/car:
    get:
      consumes:
      - application/json
      description: Get car by hsn and tsn
      parameters:
      - description: hsn
        in: query
        name: hsn
        required: true
        type: string
      - description: tsn
        in: query
        name: tsn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
      summary: Get car by hsn and tsn
      tags:
      - car
    post:
      consumes:
      - application/json
      description: Create a new car
      parameters:
      - description: Car object
        in: body
        name: Car
        required: true
        schema:
          $ref: '#/definitions/http.Car'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - car
  /api/car/{hsn}:
    get:
      consumes:
      - application/json
      description: Get car by hsn
      parameters:
      - description: hsn
        in: path
        name: hsn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get car by hsn
      tags:
      - manufacturer
  /api/car/csv:
    post:
      consumes:
      - application/json
      description: Create cars from CSV
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Create cars from CSV
      tags:
      - car
swagger: "2.0"